refactor: redistribute the vo dto classes
