+++ src/main/java/laddergame/domain/Ladder.java
@@ -1,15 +1,15 @@
 package laddergame.domain;
 
+import laddergame.vo.Position;
+
 import java.util.List;
 
 public class Ladder {
 
     private final List<Line> lines;
-    private final LadderHeight ladderHeight;
 
     public Ladder(List<Line> lines) {
         this.lines = List.copyOf(lines);
-        this.ladderHeight = new LadderHeight(lines.size());
     }
 
     public List<Line> getLadder() {
