+++ src/main/java/laddergame/domain/Line.java
@@ -1,5 +1,7 @@
 package laddergame.domain;
 
+import laddergame.vo.Position;
+
 import java.util.ArrayList;
 import java.util.List;
