+++ src/main/java/laddergame/domain/Prizes.java
@@ -1,5 +1,7 @@
 package laddergame.domain;
 
+import laddergame.vo.Position;
+
 import java.util.List;
 import java.util.stream.Collectors;
