+++ src/main/java/laddergame/vo/LadderHeight.java
@@ -1,4 +1,4 @@
-package laddergame.domain;
+package laddergame.vo;
 
 public class LadderHeight {
