+++ src/main/java/laddergame/vo/LadderLabel.java
@@ -1,4 +1,4 @@
-package laddergame.domain;
+package laddergame.vo;
 
 import laddergame.constant.ErrorCode;
