+++ src/main/java/laddergame/vo/Position.java
@@ -1,4 +1,4 @@
-package laddergame.domain;
+package laddergame.vo;
 
 import java.util.Objects;
