+++ src/test/java/laddergame/domain/LadderHeightTest.java
@@ -1,5 +1,6 @@
 package laddergame.domain;
 
+import laddergame.vo.LadderHeight;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
