+++ src/test/java/laddergame/domain/LadderMakerTest.java
@@ -1,6 +1,7 @@
 package laddergame.domain;
 
 import laddergame.util.RandomBooleanGenerator;
+import laddergame.vo.LadderHeight;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
