+++ src/test/java/laddergame/domain/LadderTest.java
@@ -1,5 +1,6 @@
 package laddergame.domain;
 
+import laddergame.vo.Position;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
