+++ src/test/java/laddergame/domain/LineTest.java
@@ -1,5 +1,6 @@
 package laddergame.domain;
 
+import laddergame.vo.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
