+++ src/test/java/laddergame/domain/PrizesTest.java
@@ -1,5 +1,6 @@
 package laddergame.domain;
 
+import laddergame.vo.Position;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -7,7 +8,6 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class PrizesTest {
